UAV/Team_Docs/Old scripts/Old_API_program/requirements_falcon.txt
falcon
transformers
torch
Pillow